fixed createedit saving average hourly price rate, user set price now saves as the full decimal pricing, not rounded to a whole number

+++ ParkHere/app/src/main/java/itp341/wang/cherrie/parkhere/CreateEditListingActivity.java
@@ -239,7 +239,7 @@ public class CreateEditListingActivity extends AppCompatActivity implements Time
             public void onClick(View v) {
                 if(!listingTitle.isEmpty() && !location.isEmpty() && !about.isEmpty() && !isPriceEditTextEmpty()){
                     // getting all values from the input fields
-                    price = priceEditText.getRawValue()/100;
+                    price = ((double)priceEditText.getRawValue())/100;
                     isTandem = tandemCheckBox.isChecked();
                     isHandicapped = handicappedCheckBox.isChecked();
                     isSUV = suvCheckBox.isChecked();
